Credential delegation support for the HTTP Handler
Support for the delegation of a credential from the client.  The handler
stores the delegated credential in the request context for consumption
by child handlers.

Also support for the credential store extension, to put the received
delegated credential into a credential cache other than the default
memory cache.

Improve test coverage